_older/django-eav-example/requirements.txt
django
django-extensions
django-jinja
django-mptt
djangorestframework
djangorestframework-filters
git+https://github.com/ArturBaybulatov/django-eav
Pillow # in case it won't install: sudo apt-get install build-essential autoconf libtool pkg-config python-opengl python-imaging python-pyrex python-pyside.qtopengl idle-python2.7 qt4-dev-tools qt4-designer libqtgui4 libqtcore4 libqt4-xml libqt4-test libqt4-script libqt4-network libqt4-dbus python-qt4 python-qt4-gl libgle3 python-dev
pydash
